Move testing helpers into their own package
Production code does not need to include testing helpers or the testing
package.
