+++ russian/russian_test.go
@@ -1,7 +1,7 @@
 package russian
 
 import (
-	"github.com/kljensen/snowball/romance"
+	"github.com/kljensen/snowball/romance/romancetest"
 	"testing"
 )
 
@@ -9,31 +9,31 @@ import (
 // or false.
 //
 func Test_isLowerVowel(t *testing.T) {
-	testCases := []romance.WordBoolTestCase{
+	testCases := []romancetest.WordBoolTestCase{
 		// These are all vowels.
 		{"аеиоуыэюя", true},
 		// None of these are vowels.
 		{"бвгджзйклмнпрстфхцчшщъь", false},
 	}
-	romance.RunRunewiseBoolTest(t, isLowerVowel, testCases)
+	romancetest.RunRunewiseBoolTest(t, isLowerVowel, testCases)
 }
 
 // Test stopWords for things we know should be true
 // or false.
 //
 func Test_stopWords(t *testing.T) {
-	testCases := []romance.WordBoolTestCase{
+	testCases := []romancetest.WordBoolTestCase{
 		{"была", true},
 		{"нас", true},
 		{"меня", true},
 		{"химическое", false},
 		{"машиностроение", false},
 	}
-	romance.RunWordBoolTest(t, isStopWord, testCases)
+	romancetest.RunWordBoolTest(t, isStopWord, testCases)
 }
 
 func Test_findRegions(t *testing.T) {
-	testCases := []romance.FindRegionsTestCase{
+	testCases := []romancetest.FindRegionsTestCase{
 		{"кистей", 3, 6, 2},
 		{"пугает", 3, 6, 2},
 		{"горячей", 3, 5, 2},
@@ -136,13 +136,13 @@ func Test_findRegions(t *testing.T) {
 		{"полярный", 3, 5, 2},
 	}
 
-	romance.RunFindRegionsTest(t, findRegions, testCases)
+	romancetest.RunFindRegionsTest(t, findRegions, testCases)
 }
 
 // Test step1
 //
 func Test_step1(t *testing.T) {
-	testCases := []romance.StepTestCase{
+	testCases := []romancetest.StepTestCase{
 		{"оснеженные", 2, 5, 1, true, "оснеженн", 2, 5, 1},
 		{"умираю", 2, 4, 1, true, "умира", 2, 4, 1},
 		{"старосте", 4, 6, 3, true, "старост", 4, 6, 3},
@@ -295,14 +295,14 @@ func Test_step1(t *testing.T) {
 		{"овчинин", 2, 5, 1, false, "овчинин", 2, 5, 1},
 		{"ков", 3, 3, 2, false, "ков", 3, 3, 2},
 	}
-	romance.RunStepTest(t, step1, testCases)
+	romancetest.RunStepTest(t, step1, testCases)
 
 }
 
 // Test step2
 //
 func Test_step2(t *testing.T) {
-	testCases := []romance.StepTestCase{
+	testCases := []romancetest.StepTestCase{
 		{"незнани", 3, 6, 2, true, "незнан", 3, 6, 2},
 		{"однообрази", 2, 6, 1, true, "однообраз", 2, 6, 1},
 		{"развити", 3, 6, 2, true, "развит", 3, 6, 2},
@@ -405,14 +405,14 @@ func Test_step2(t *testing.T) {
 		{"додума", 3, 5, 2, false, "додума", 3, 5, 2},
 		{"додума", 3, 5, 2, false, "додума", 3, 5, 2},
 	}
-	romance.RunStepTest(t, step2, testCases)
+	romancetest.RunStepTest(t, step2, testCases)
 
 }
 
 // Test step3
 //
 func Test_step3(t *testing.T) {
-	testCases := []romance.StepTestCase{
+	testCases := []romancetest.StepTestCase{
 		{"подробност", 3, 6, 2, true, "подробн", 3, 6, 2},
 		{"подробност", 3, 6, 2, true, "подробн", 3, 6, 2},
 		{"подробност", 3, 6, 2, true, "подробн", 3, 6, 2},
@@ -515,14 +515,14 @@ func Test_step3(t *testing.T) {
 		{"арестант", 2, 4, 1, false, "арестант", 2, 4, 1},
 		{"шепнул", 3, 6, 2, false, "шепнул", 3, 6, 2},
 	}
-	romance.RunStepTest(t, step3, testCases)
+	romancetest.RunStepTest(t, step3, testCases)
 
 }
 
 // Test step4
 //
 func Test_step4(t *testing.T) {
-	testCases := []romance.StepTestCase{
+	testCases := []romancetest.StepTestCase{
 		{"раскольничь", 3, 6, 2, true, "раскольнич", 3, 6, 2},
 		{"создань", 3, 6, 2, true, "создан", 3, 6, 2},
 		{"доверенн", 3, 5, 2, true, "доверен", 3, 5, 2},
@@ -625,7 +625,7 @@ func Test_step4(t *testing.T) {
 		{"цепля", 3, 5, 2, false, "цепля", 3, 5, 2},
 		{"дворянин", 4, 6, 3, false, "дворянин", 4, 6, 3},
 	}
-	romance.RunStepTest(t, step4, testCases)
+	romancetest.RunStepTest(t, step4, testCases)
 }
 
 // Test a large set of words for which we know
